move ColorPool out
